requirements.txt
streamlitnumpypillowopencv-python-headlessscikit-imagescipypdf2imagepytesseract